Split FastBitmap between files by making it a partial class.
